refactor(registry): split config module
